sovled issues with some tests afer restructure, good to go
