+++ web/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./App.css";
-import AppRouter from "./components/AppRouter";
+import AppRouter from "./AppRouter/AppRouter";
 import Header from "./Global/Header/Header";
 
 function App() {
