+++ web/src/AppRouter/AppRouter.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import HomePage from "./HomePage";
+import HomePage from "../Pages/HomePage/HomePage";
 import TopicPage from "../Pages/TopicPage/TopicPage";
 
 export default function AppRouter() {
