+++ web/src/Global/Header/Header.jsx
@@ -1,6 +1,5 @@
 /* eslint-disable react/jsx-no-bind */
 import React, { useState, useEffect } from "react";
-import explainailogo from "../../images/explainai-logo.png";
 import "./Header.scss";
 import DropDown from "./DropDown";
 import ButtonSelector from "./ButtonSelector";
@@ -22,7 +21,7 @@ export default function Header() {
   return (
     <div>
       <header>
-        <img src={explainailogo} alt="logo" />
+        <img src="./explainai-logo.png" alt="logo" />
         <div>
           <p>{labelText}</p>
           <ButtonSelector age={age} changeAge={changeAge} />
