+++ web/src/Pages/TopicPage/BreadCrumbs/Breadcrumbs.jsx
@@ -1,7 +1,8 @@
+/* eslint-disable react/button-has-type */
 import { Link } from "react-router-dom";
 import { React } from "react";
 
-import "../Styles/TopicBreadcrumbs/TopicBreadcrumbs.css";
+import "../../../Styles/TopicBreadcrumbs/TopicBreadcrumbs.css";
 
 export default function Breadcrumbs({ grandParent, parent }) {
   return (
@@ -18,10 +19,9 @@ export default function Breadcrumbs({ grandParent, parent }) {
           </span>
         </span>
       </nav>
-
       <div id="mobile" data-testid="mobile">
         <Link data-testid="mobile-grandParent" to={`/${grandParent.slug}`}>
-          <button type="button" className="mobile-parent-btn">
+          <button className="mobile-parent-btn">
             <span className="ml-7">
               {" "}
               {grandParent.name}
@@ -31,7 +31,7 @@ export default function Breadcrumbs({ grandParent, parent }) {
         </Link>
 
         <Link data-testid="mobile-parent" to={`/${parent.slug}`}>
-          <button type="button" className="mobile-grandParent-btn">
+          <button className="mobile-grandParent-btn">
             <span className="ml-7">
               {" "}
               {parent.name}
