+++ web/src/Pages/TopicPage/TopicCard/TopicCard.test.js
@@ -1,13 +1,10 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import TopicCard from "../Pages/TopicPage/TopicCard";
+import TopicCard from "./TopicCard";
 import testData from "./dummy_data.json";
 
 const topic = testData.topic[0];
 describe("Tests for TopicCard using RTL", () => {
-  beforeEach(() => {
-    // topic = { ...testData.topic[0] }; // reset topic
-  });
   test("Renders everything in the right roles", () => {
     render(<TopicCard topic={topic} audience={10} />);
     const title = screen.getByRole("heading");
@@ -19,14 +16,8 @@ describe("Tests for TopicCard using RTL", () => {
     expect(image.src).toBe(topic.image);
   });
   test("Renders dummy image if dummy data has no image", () => {
-    // topic.image = false;
     render(<TopicCard topic={{ ...topic, image: false }} audience={10} />);
     const image = screen.getByRole("img");
     expect(image.src).toBe("http://localhost/no-image.jpeg");
   });
-  // test("JSON is not the correct structure, so an error box appears", () => {
-  //   // topic.descriptions = "";
-  //   render(<TopicCard topic={{ ...topic, descriptions: "" }} audience={20} />);
-  //   expect(screen.getByTestId("error-message")).toBeInTheDocument();
-  // });
 });
