+++ web/src/Styles/TopicCard/_variables.scss
