+++ web/src/styles/TopicBreadcrumbs/TopicBreadcrumbs.css
@@ -1,15 +1,26 @@
+@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@200&family=Tilt+Warp&display=swap");
 @media (max-width: 400px) {
   #desktop {
     display: none;
   }
 }
-
+@media (min-width: 400px) {
+  #desktop {
+    display: block;
+  }
+}
 @media (min-width: 1000px) {
   #desktop {
+    display: block;
     margin-left: 15%;
   }
 }
 
+@media (max-width: 400px) {
+  #mobile {
+    display: block;
+  }
+}
 @media (min-width: 400px) {
   #mobile {
     display: none;
@@ -35,4 +46,4 @@
 
 .mobile-grandParent-btn {
   @apply bg-backdrop pb-2 pt-2 text-left text-white capitalize w-screen;
-} /*# sourceMappingURL=TopicBreadcrumbs.css.map */
+}/*# sourceMappingURL=TopicBreadcrumbs.css.map */
