--- web/src/styles/variables.css
@@ -1,1 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@200&family=Tilt+Warp&display=swap");/*# sourceMappingURL=variables.css.map */
