+++ web/src/utils/networking.js
@@ -4,7 +4,6 @@ export default async function fetchData(url) {
     const apiResponseJSON = await apiResponse.json();
     return apiResponseJSON;
   } catch (e) {
-    console.log("unable to fetch data", e);
     return "unable to fetch data";
   }
 }
